# openapi/ingestion-trusted-human-v1.yaml
openapi: 3.1.0
info: {title: OUF Ingestion Runtime Trusted Human API, version: 1.0.0}
servers: [{url: /api/trusted-human/v1/ingestion}]
paths:
  /issues:
    get:
      operationId: searchIngestionRuntimeIssues
      x-ouf-mcp-tool-eligible: true
      x-ouf-authorization-capability: ouf.ingestion.issue.search
      parameters:
        - {in: query, name: state, schema: {type: string, enum: [OPEN, RESOLVED, DISMISSED]}}
        - {in: query, name: severity, schema: {type: string, enum: [WARNING, ERROR]}}
        - {in: query, name: sourceId, schema: {type: string}}
        - {in: query, name: runId, schema: {type: string, format: uuid}}
        - {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 200}}
      responses: {'200': {description: Tenant-scoped runtime issues without raw payload}, '403': {description: Forbidden}}
  /issues/schema-observations:
    get:
      operationId: searchIngestionSchemaObservations
      x-ouf-mcp-tool-eligible: true
      x-ouf-authorization-capability: ouf.ingestion.issue.search
      parameters:
        - {in: query, name: sourceId, schema: {type: string}}
        - {in: query, name: outcome, schema: {type: string, enum: [MATCH, COMPATIBLE_ADDITIVE, UNKNOWN_VERSION, INCOMPATIBLE]}}
        - {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 200}}
      responses: {'200': {description: Persisted schema observations}, '403': {description: Forbidden}}
  /issues/{id}:
    get:
      operationId: getIngestionRuntimeIssue
      x-ouf-mcp-tool-eligible: true
      x-ouf-authorization-capability: ouf.ingestion.issue.read
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: Issue detail and safe evidence references}, '404': {description: Not found in actor tenant}}
  /issues/{id}/explain:
    get:
      operationId: explainIngestionRuntimeIssue
      x-ouf-mcp-tool-eligible: true
      x-ouf-authorization-capability: ouf.ingestion.issue.explain
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: Deterministic safe explanation and recommended action}, '404': {description: Not found in actor tenant}}
  /issues/{id}/resolve:
    post:
      operationId: resolveIngestionRuntimeIssue
      x-ouf-mcp-tool-eligible: true
      x-ouf-requires-human-user: true
      x-ouf-authorization-capability: ouf.ingestion.issue.resolve
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/IssueCloseRequest'}}}}
      responses: {'204': {description: Resolved and audited}, '409': {description: Optimistic lock or state conflict}}
  /issues/{id}/dismiss:
    post:
      operationId: dismissIngestionRuntimeIssue
      x-ouf-mcp-tool-eligible: true
      x-ouf-requires-human-user: true
      x-ouf-authorization-capability: ouf.ingestion.issue.dismiss
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/IssueCloseRequest'}}}}
      responses: {'204': {description: Dismissed with reason and audit}, '409': {description: Optimistic lock or state conflict}}
  /issues/{id}/onboarding-review:
    post:
      operationId: requestOnboardingReviewForIssue
      x-ouf-mcp-tool-eligible: true
      x-ouf-authorization-capability: ouf.ingestion.onboarding-review.request
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ReviewRequest'}}}}
      responses: {'202': {description: Durable review request queued}, '409': {description: State conflict}}
  /quarantine:
    get:
      operationId: searchIngestionQuarantine
      x-ouf-mcp-tool-eligible: false
      x-ouf-authorization-capability: ingestion.quarantine.read
      parameters:
        - {in: query, name: state, schema: {type: string, enum: [OPEN, RETRY_READY, REPROCESSING, RESOLVED, DISMISSED, SUPERSEDED]}}
        - {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 200}}
      responses: {'200': {description: Redacted quarantine list}, '403': {description: Trusted HUMAN_USER context or capability missing}}
  /quarantine/{id}:
    get:
      operationId: getIngestionQuarantine
      x-ouf-mcp-tool-eligible: false
      x-ouf-authorization-capability: ingestion.quarantine.read
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: Redacted quarantine detail}, '403': {description: Forbidden}}
  /quarantine/{id}/replay:
    post:
      operationId: replayIngestionQuarantine
      x-ouf-mcp-tool-eligible: false
      x-ouf-authorization-capability: ingestion.quarantine.replay
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ReplayRequest'}}}
      responses: {'202': {description: Replay request durably queued}, '403': {description: Forbidden}, '409': {description: Item is not open}}
  /quarantine/{id}/preview:
    get:
      operationId: inspectIngestionQuarantineSafely
      x-ouf-mcp-tool-eligible: true
      x-ouf-authorization-capability: ouf.ingestion.quarantine.inspect
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: Tenant-scoped preview with object fingerprint and no raw payload}, '404': {description: Not found in actor tenant}}
  /quarantine/{id}/reprocess:
    post:
      operationId: reprocessIngestionQuarantine
      x-ouf-mcp-tool-eligible: true
      x-ouf-requires-human-user: true
      x-ouf-authorization-capability: ouf.ingestion.quarantine.reprocess
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ReplayRequest'}}}}
      responses: {'202': {description: Durable reprocessing queued}, '409': {description: Lifecycle conflict}}
  /quarantine/{id}/dismiss:
    post:
      operationId: dismissIngestionQuarantine
      x-ouf-mcp-tool-eligible: true
      x-ouf-requires-human-user: true
      x-ouf-authorization-capability: ouf.ingestion.issue.dismiss
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/QuarantineCloseRequest'}}}}
      responses: {'204': {description: Dismissed and audited}, '409': {description: Lifecycle conflict}}
  /quarantine/{id}/supersede:
    post:
      operationId: supersedeIngestionQuarantine
      x-ouf-mcp-tool-eligible: true
      x-ouf-requires-human-user: true
      x-ouf-authorization-capability: ouf.ingestion.issue.dismiss
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/QuarantineCloseRequest'}}}}
      responses: {'204': {description: Superseded and audited}, '409': {description: Lifecycle conflict}}
  /quarantine/replays/{replayId}:
    get:
      operationId: getIngestionReplay
      x-ouf-mcp-tool-eligible: false
      x-ouf-authorization-capability: ingestion.replay.read
      parameters:
        - {in: path, name: replayId, required: true, schema: {type: string, format: uuid}}
      responses: {'200': {description: Replay execution status without payload or downstream receipt}, '403': {description: Forbidden}}
  /logs/search:
    post:
      operationId: searchProtectedIngestionLogs
      x-ouf-mcp-tool-eligible: false
      x-ouf-authorization-capability: ingestion.log.read
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LogQuery'}}}}
      responses: {'200': {description: Bounded and server-redacted operational events}, '403': {description: Forbidden}}
  /logs/aggregate:
    post:
      operationId: aggregateProtectedIngestionLogs
      x-ouf-mcp-tool-eligible: false
      x-ouf-authorization-capability: ingestion.log.aggregate
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [query, dimension], properties: {query: {$ref: '#/components/schemas/LogQuery'}, dimension: {type: string, enum: [severity, event_type, source_id]}}}}}}
      responses: {'200': {description: Bounded server-side aggregation}, '403': {description: Forbidden}}
components:
  parameters:
    Id: {in: path, name: id, required: true, schema: {type: string, format: uuid}}
  schemas:
    IssueCloseRequest:
      type: object
      required: [reason, expectedVersion]
      properties: {reason: {type: string, minLength: 1, maxLength: 1000}, expectedVersion: {type: integer, minimum: 0}}
      additionalProperties: false
    ReviewRequest:
      type: object
      required: [expectedVersion]
      properties: {expectedVersion: {type: integer, minimum: 0}}
      additionalProperties: false
    ReplayRequest:
      type: object
      required: [targetBundleRef]
      properties: {targetBundleRef: {type: string, minLength: 1}}
      additionalProperties: false
    QuarantineCloseRequest:
      type: object
      required: [reason, expectedVersion]
      properties:
        reason: {type: string, minLength: 1, maxLength: 1000}
        supersededBy: {type: string, minLength: 1}
        expectedVersion: {type: integer, minimum: 0}
      additionalProperties: false
    LogQuery:
      type: object
      required: [from, to, purpose]
      properties:
        from: {type: string, format: date-time}
        to: {type: string, format: date-time}
        severity: {type: string, enum: [INFO, WARN, ERROR]}
        correlationId: {type: string}
        runId: {type: string, format: uuid}
        sourceId: {type: string}
        limit: {type: integer, minimum: 1, maximum: 200}
        purpose: {type: string, minLength: 1, maxLength: 500}
      additionalProperties: false